+++ src/app/components/forms/board-edit/board-edit-form.component.ts
@@ -1,4 +1,4 @@
-import { Component, EventEmitter, OnInit, Output } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { TodoBoard } from '../../../models/todo.model';
 
 @Component({
@@ -8,37 +8,56 @@ import { TodoBoard } from '../../../models/todo.model';
 })
 export class BoardEditFormComponent implements OnInit {
 
-    @Output() closeForm: EventEmitter<void> = new EventEmitter<void>();
-    @Output() saveClick: EventEmitter<TodoBoard> = new EventEmitter<TodoBoard>();
+    @Input() todoBoard: TodoBoard = null;
 
+    @Output() clickedClose: EventEmitter<void> = new EventEmitter<void>();
+    @Output() clickedSave: EventEmitter<TodoBoard> = new EventEmitter<TodoBoard>();
+
+    public createdDate: Date = null;
     public description: string = '';
+    public isReadOnly: boolean = false;
     public name: string = '';
 
     constructor() {}
 
-    ngOnInit(): void {}
+    ngOnInit(): void {
+        if (!!this.todoBoard) {
+            const { createdDate, description, isArchived, name } = this.todoBoard;
+            this.createdDate = createdDate;
+            this.description = description;
+            this.isReadOnly = isArchived;
+            this.name = name;
+        }
+    }
 
-    public handleChangeDescription(event: any): void {
+    public handleChangedDescription(event: any): void {
         if (!!event && !!event.target) {
             this.description = event.target.value;
         }
     }
 
-    public handleChangeName(event: any): void {
+    public handleChangedName(event: any): void {
         if (!!event && !!event.target) {
             this.name = event.target.value;
         }
     }
 
-    public handleClickButtonSave(): void {
+    public handleClickedSave(): void {
         if (!!this.name) {
             const id: string = 'new-board-id-' + new Date().getTime();
-            this.saveClick.emit(new TodoBoard(id, this.name, [], this.description));
-            this.closeForm.emit();
+            const board: TodoBoard = new TodoBoard(id, this.name, [], this.description, true, false, new Date());
+            if (!!this.todoBoard) {
+                board.id = this.todoBoard.id;
+                board.isArchived = this.todoBoard.isArchived;
+                board.isOpen = this.todoBoard.isOpen;
+                board.tasks = this.todoBoard.tasks;
+            }
+            this.clickedSave.emit(board);
+            this.clickedClose.emit();
         }
     }
 
-    public handleClickButtonCancel(): void {
-        this.closeForm.emit();
+    public handleClickedCancel(): void {
+        this.clickedClose.emit();
     }
 }